refactor: 萃取 CacheInterface 統一快取介面階層
- 新增 CacheInterface 包含 7 個核心方法（get/set/has/delete/clear/remember/getStats）
- CacheServiceInterface 改為 extends CacheInterface，保留批次操作方法
- CacheManagerInterface 改為 extends CacheInterface，保留管理操作方法
- CacheService 更新 get() 與 remember() 簽章以符合新介面
- CacheManager 與 PrefixedCacheManager 無需修改（簽章已相容）
- 176 tests pass, PHPStan Level 10: 0 errors
